tools/image-pixel-processor/requirements.txt
Pillow>=10.4.0
numpy>=2.0.0
rembg[cpu]>=2.0.59
opencv-python>=4.10.0
